# Conflicts:
#	docker/hvd/Dockerfile.hvd-apex
#	docker/hvd/Dockerfile.hvd-base
#	docker/main/Dockerfile.apex
#	docker/msdp/Dockerfile.msdp-apex

+++ docker/hvd/build_all.sh
@@ -2,6 +2,21 @@
 
 echo "Build all Horovod flavoured PyTorch-Ignite images"
 
+retry()
+{
+    cmd=$1 msg=$2
+    counter=0 limit=3
+    while [ "$counter" -lt "$limit" ]; do
+        echo "(Re-)Try: $cmd"
+        bash -c "$cmd" && break
+        echo $msg
+        counter="$(( $counter + 1 ))"
+    done
+    if [ $counter -eq $limit ]; then
+        exit 1
+    fi
+}
+
 # Start script from ignite docker folder
 if [ ! -d hvd ]; then
     echo "Can not find 'hvd' folder"
@@ -12,18 +27,19 @@ fi
 curr_dir=$PWD
 cd $curr_dir/hvd
 
-set -xeu
+set -eu
 
-image_name="hvd-base"
+image_tag=""
+pth_version=${PTH_VERSION:-1.6.0-cuda10.1-cudnn7}
+hvd_version=${HVD_VERSION:-v0.21.0}
 
-docker build -t pytorchignite/${image_name}:latest -f Dockerfile.${image_name} .
-image_tag=`docker run --rm -i pytorchignite/${image_name}:latest python -c "import torch; import ignite; print(torch.__version__ + \"-\" + ignite.__version__, end=\"\")"`
-docker tag pytorchignite/${image_name}:latest pytorchignite/${image_name}:${image_tag}
-
-for image_name in "hvd-vision" "hvd-nlp" "hvd-apex" "hvd-apex-vision" "hvd-apex-nlp"
+for image_name in "hvd-base" "hvd-vision" "hvd-nlp" "hvd-apex" "hvd-apex-vision" "hvd-apex-nlp"
 do
 
-    docker build -t pytorchignite/${image_name}:latest -f Dockerfile.${image_name} .
+    retry "docker build --build-arg PTH_VERSION=${pth_version} --build-arg HVD_VERSION=${hvd_version} -t pytorchignite/${image_name}:latest -f Dockerfile.${image_name} ." "\nBuild failed: ${image_name}"
+    if [ -z $image_tag ]; then
+        image_tag=`docker run --rm -i pytorchignite/${image_name}:latest python -c "import torch; import ignite; print(torch.__version__ + \"-\" + ignite.__version__, end=\"\")"`
+    fi
     docker tag pytorchignite/${image_name}:latest pytorchignite/${image_name}:${image_tag}
 
 done
